Added CredentialProviderSelector protocol
This is a credential provider which has to select an internal credential provider. Use the protocol with the RuntimeSelectorCredentialProvider.
