+++ Sources/SotoCore/Credential/RuntimeSelectorCredentialProvider.swift
@@ -20,37 +20,18 @@ import SotoSignerV4
 /// get credentials from a list of possible credential providers. Goes through list of providers from start to end
 /// attempting to get credentials. Once it finds a `CredentialProvider` that supplies credentials use that
 /// one
-class RuntimeSelectorCredentialProvider: CredentialProvider {
-    /// the provider chosen to supply credentials
-    var internalProvider: CredentialProvider? {
-        self.lock.withLock {
-            _internalProvider
-        }
-    }
-
+class RuntimeSelectorCredentialProvider: CredentialProviderSelector {
     /// promise to find a credential provider
     let startupPromise: EventLoopPromise<CredentialProvider>
-
-    private let lock = Lock()
-    private var _internalProvider: CredentialProvider?
+    let lock = Lock()
+    var _internalProvider: CredentialProvider?
 
     init(providers: [CredentialProviderFactory], context: CredentialProviderFactory.Context) {
         self.startupPromise = context.eventLoop.makePromise(of: CredentialProvider.self)
+        self.startupPromise.futureResult.whenSuccess { result in
+            self.internalProvider = result
+        }
         self.setupInternalProvider(providers: providers, context: context)
-    }
-
-    func shudown(on eventLoop: EventLoop) -> EventLoopFuture<Void> {
-        return self.startupPromise.futureResult.map { _ in }.hop(to: eventLoop)
-    }
-
-    func getCredential(on eventLoop: EventLoop, logger: Logger) -> EventLoopFuture<Credential> {
-        if let provider = internalProvider {
-            return provider.getCredential(on: eventLoop, logger: logger)
-        }
-
-        return self.startupPromise.futureResult.hop(to: eventLoop).flatMap { provider in
-            return provider.getCredential(on: eventLoop, logger: logger)
-        }
     }
 
     /// goes through list of providers. If provider is able to provide credentials then use that one, otherwise move onto the next
@@ -67,7 +48,6 @@ class RuntimeSelectorCredentialProvider: CredentialProvider {
                 switch result {
                 case .success:
                     context.logger.info("Select credential provider", metadata: ["aws-credential-provider": .string("\(provider)")])
-                    self._internalProvider = provider
                     self.startupPromise.succeed(provider)
                 case .failure:
                     _setupInternalProvider(index + 1)
